+++ src/background/badge.js
