+++ src/background/clipboard-access.js
