+++ src/background/message-handler.js
@@ -1,6 +1,6 @@
 import * as BrowserAsMarkdown from './browser-as-markdown.js';
 import copy from './clipboard-access.js';
-import { flashSuccessBadge } from '../../lib/badge.js';
+import { flashSuccessBadge } from './badge.js';
 
 async function handleCopy(action) {
   /** @type {string} */
